pyproject.toml
[project]
name = "game-recording"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dotenv>=0.9.9",
    "mss>=10.1.0",
    "pillow>=12.1.1",
    "psutil>=7.2.2",
    "pydirectinput>=1.0.4",
    "pynput>=1.8.1",
    "pywin32>=311",
    "requests>=2.32.5",
    "schedule>=1.2.2",
]
